+++ backend/pkg/api/fixtures.go
@@ -1,0 +1,6 @@
+package api
+
+type Response_New struct {
+	Weight string `json:"weight"`
+	Force  bool   `json:"force"`
+}
