+++ backend/pkg/db/fixtures.go
@@ -1,0 +1,6 @@
+package db
+
+type Entry struct {
+	Date   string `json:"date"`
+	Weight string `json:"weight"`
+}
