+++ docker-compose.yaml
@@ -2,6 +2,10 @@ version: "3.9"
 services:
   backend:
     build: backend/.
+    environment:
+      - DB_USER
+      - DB_PASSWORD
+      - DB_PORT
     ports:
       - "10000:10000"
   frontend:
